# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "단기 팝업스토어 턴키 예약·무인 운영 플랫폼 API. 모든 응답은 {data, error} 봉투로 감싸진다.",
    "title": "PopupReady API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "로컬 개발"
    }
  ],
...
  "paths": {
...
    "/api/v1/contracts/{id}": {
      "get": {
        "description": "조항 전문·서명 시각·무결성 해시를 돌려준다. 분쟁 시 소명 자료 경로다. 당사자에게만 열린다.",
        "operationId": "getContract",
        "parameters": [
          {
            "description": "계약 ID",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContractResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 값 검증 실패"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "인증이 필요함"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "역할 또는 당사자 자격이 없음"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "리소스를 찾을 수 없음"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "계약 열람",
        "tags": [
          "contract"
        ]
      }
    },
    "/api/v1/contracts/{id}/sign": {
      "post": {
        "description": "로그인 사용자가 해당 계약의 당사자인지 확인한 뒤 서명 시각을 기록한다. 양측이 모두 서명하면 계약은 SIGNED, 예약 요청은 CONTRACT_SIGNED가 된다. 당사자가 아니면 403, 이미 서명했으면 409다.",
        "operationId": "signContract",
        "parameters": [
          {
            "description": "계약 ID",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContractResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 값 검증 실패"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "인증이 필요함"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "역할 또는 당사자 자격이 없음"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "리소스를 찾을 수 없음"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "전자 서명",
        "tags": [
          "contract"
        ]
      }
    },
...
    "/api/v1/reservation-requests/{id}/contract": {
      "get": {
        "description": "예약 요청에 딸린 계약을 가져온다. 아직 없으면 404다. 빌더에서 계약 단계로 재진입할 때(새로고침·서명 링크 재방문) 계약 ID를 모르는 채로 기존 계약을 되찾기 위한 경로다. 당사자에게만 열린다.",
        "operationId": "getContractByReservation",
        "parameters": [
          {
            "description": "예약 요청 ID",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContractResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 값 검증 실패"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "인증이 필요함"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "역할 또는 당사자 자격이 없음"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "리소스를 찾을 수 없음"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "예약 요청의 계약 조회",
        "tags": [
          "contract"
        ]
      },
      "post": {
        "description": "예약 요청 데이터를 표준 템플릿에 바인딩해 조항 전문을 스냅샷으로 저장한다. 예약 요청 상태는 CONTRACT_PENDING으로 전이된다. 예약 하나에 계약은 하나이며, 이미 있으면 409다 — 그때는 조회 API로 기존 계약을 가져간다. 브랜드·건물주 당사자만 부를 수 있으며 그 외에는 403이다.",
        "operationId": "createContract",
        "parameters": [
          {
            "description": "예약 요청 ID",
            "example": 1,
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContractResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "요청 값 검증 실패"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "인증이 필요함"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "역할 또는 당사자 자격이 없음"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "리소스를 찾을 수 없음"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "서버 오류"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "계약서 생성",
        "tags": [
          "contract"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "에러 상세. 성공 응답에서는 null이다.",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드 상수. 클라이언트는 이 값으로 분기한다",
            "enum": [
              "VALIDATION_FAILED",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "UNSUPPORTED_MEDIA_TYPE",
              "INTERNAL_ERROR",
              "EMAIL_ALREADY_EXISTS",
              "INVALID_CREDENTIALS",
              "SPACE_NOT_FOUND",
              "FIXTURE_NOT_FOUND",
              "RESERVATION_REQUEST_NOT_FOUND",
              "LAYOUT_OUT_OF_BOUNDS",
              "LAYOUT_OVERLAP",
              "FIXTURE_STOCK_EXCEEDED",
              "FIXTURE_UNAVAILABLE",
              "POWER_LIMIT_EXCEEDED",
              "CONTRACT_NOT_FOUND",
              "NOT_CONTRACT_PARTY",
              "CONTRACT_ALREADY_EXISTS",
              "CONTRACT_ALREADY_SIGNED",
              "CONTRACT_INTEGRITY_VIOLATION",
              "REFRESH_TOKEN_INVALID",
              "PAYMENT_ALREADY_COMPLETED",
              "PAYMENT_AMOUNT_MISMATCH",
              "LOCK_ACQUISITION_FAILED",
              "DOOR_NOT_YET_OPENABLE"
            ],
            "example": "VALIDATION_FAILED"
          },
          "message": {
            "type": "string",
            "description": "사람이 읽는 설명. 분기 조건으로 쓰지 말 것",
            "example": "email은 필수입니다"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "실패 응답 봉투. data는 항상 null이고 error에 코드·메시지가 담긴다.",
        "properties": {
          "data": {
            "description": "실패 시 항상 null"
          },
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "data",
          "error"
        ]
      },
...
      "ApiResponseContractResponse": {
        "type": "object",
        "description": "공통 응답 봉투",
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContractResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "성공 시 페이로드. 실패 시 null"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ],
            "description": "실패 시 에러 상세. 성공 시 null"
          }
        },
        "required": [
          "data",
          "error"
        ]
      },
...
      "ClauseDto": {
        "type": "object",
        "description": "계약 조항",
        "properties": {
          "body": {
            "type": "string",
            "description": "조항 전문",
            "example": "본 계약은 팝업스토어 단기 운영을 목적으로 한다."
          },
          "title": {
            "type": "string",
            "description": "조항 제목",
            "example": "제1조 (목적)"
          }
        },
        "required": [
          "body",
          "title"
        ]
      },
      "ContractResponse": {
        "type": "object",
        "description": "계약",
        "properties": {
          "brandSignedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "브랜드 서명 시각(UTC). 미서명이면 null",
            "example": "2026-08-22T05:12:31Z"
          },
          "clauses": {
            "type": "array",
            "description": "생성 시점의 조항 전문 스냅샷",
            "items": {
              "$ref": "#/components/schemas/ClauseDto"
            }
          },
          "contentHash": {
            "type": "string",
            "description": "조항 전문+타임스탬프의 SHA-256 무결성 해시",
            "example": "3f2b7c1d9a4e5f60718293a4b5c6d7e8f90a1b2c3d4e5f60718293a4b5c6d7e8"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "계약 ID",
            "example": 1
          },
          "landlordSignedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "건물주 서명 시각(UTC). 미서명이면 null",
            "example": "2026-08-22T06:40:02Z"
          },
          "reservationRequestId": {
            "type": "integer",
            "format": "int64",
            "description": "대상 예약 요청 ID",
            "example": 1
          },
          "status": {
            "type": "string",
            "description": "계약 상태",
            "enum": [
              "PENDING",
              "SIGNED"
            ]
          },
          "templateVersion": {
            "type": "string",
            "description": "템플릿 버전",
            "example": "v1"
          },
          "title": {
            "type": "string",
            "description": "계약 명칭. 일시사용 임대차 요건 보존을 위해 고정된 값이다",
            "example": "단기 공간사용 제휴계약"
          }
        },
        "required": [
          "brandSignedAt",
          "clauses",
          "contentHash",
          "id",
          "landlordSignedAt",
          "reservationRequestId",
          "status",
          "templateVersion",
          "title"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "로그인 응답의 accessToken을 Bearer로 전달",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}